Finish Class Trainer implementation
- Add new field reqClass in npc_trainer
- Implement IsTalentSpell function
- Make players start with no abilities on action bars and no starter gear
- Transmog cleanup

+++ src/server/game/Entities/Creature/Creature.h
@@ -385,7 +385,7 @@ struct TrainerSpell
     uint16 reqSkill;
     uint16 reqSkillValue;
     uint8 reqLevel;
-    uint8 reqClass;
+    uint16 reqClass;
     uint32 learnedSpell[3];
 
     // helpers
